Revert "update Darts.java score method signature to match the parameter types used in DartsTest.java"
This reverts commit b52f3433d4eadc14b2732adfe9327849bf211d11.

+++ exercises/practice/darts/src/main/java/Darts.java
@@ -1,5 +1,5 @@
 class Darts {
-    int score(double xOfDart, double yOfDart) {
+    int score(int xOfDart, int yOfDart) {
         throw new UnsupportedOperationException("Delete this statement and write your own implementation.");
     }
 }
